requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser
Pillow==9.0.0
django-filter==23.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
